ch12/ch12-microservices-interop/requirements.txt
a2wsgi>=1.10.4
alembic>=1.13.1
aniso8601>=7.0.0
annotated-types>=0.6.0
anyio>=4.3.0
apispec>=6.6.1
ariadne>=0.23.0
asgiref>=3.8.1
blinker>=1.7.0
click>=8.1.7
colorama>=0.4.6
Deprecated>=1.2.14
Django>=5.0.4
django-cors-headers>=4.3.1
django-rest-framework>=0.1.0
djangorestframework>=3.15.1
fastapi>=0.110.2
Flask>=3.0.3
Flask-Cors>=4.0.0
Flask-HTTPAuth>=4.8.0
flask-marshmallow>=1.2.1
Flask-Migrate>=4.0.7
flask-openapi3>=3.1.1
Flask-SQLAlchemy>=3.1.1
googleapis-common-protos>=1.59.1
graphql-core>=3.2.3
graphql-relay>=2.0.1
greenlet>=3.0.3
grpcio>=1.63.0
h11>=0.14.0
idna>=3.7
importlib-metadata>=7.0.0
itsdangerous>=2.2.0
Jinja2>=3.1.3
Mako>=1.3.3
MarkupSafe>=2.1.5
marshmallow>=3.21.2
opentelemetry-api>=1.24.0
opentelemetry-exporter-jaeger>=1.21.0
opentelemetry-exporter-jaeger-proto-grpc>=1.21.0
opentelemetry-exporter-jaeger-thrift>=1.21.0
opentelemetry-instrumentation>=0.45b0
opentelemetry-instrumentation-flask>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0
opentelemetry-instrumentation-wsgi>=0.45b0
opentelemetry-sdk>=1.24.0
opentelemetry-semantic-conventions>=0.45b0
opentelemetry-util-http>=0.45b0
packaging>=24.0
peewee>=3.17.3
peewee-migrations>=0.3.32
promise>=2.3
protobuf>=4.25.3
psycopg2>=2.9.9
pydantic>=2.7.1
pydantic_core>=2.18.2
Rx>=1.6.3
singledispatch>=3.7.0
six>=1.16.0
sniffio>=1.3.1
SQLAlchemy>=2.0.29
sqlparse>=0.5.0
starlette>=0.37.2
thrift>=0.20.0
toml>=0.10.2
tornado>=6.4
typing_extensions>=4.11.0
tzdata>=2024.1
uvicorn>=0.29.0
webargs>=8.4.0
Werkzeug>=3.0.2
wrapt>=1.16.0
zipp>=3.18.1
